ci: Build, deploy, and manage Docker image with GitHub Actions (#8)
This workflow automates the process of building, deploying, and managing Docker images for the application. It triggers on push to the main branch and on pull requests. The workflow sets environment variables, builds and pushes the Docker image to Docker Hub, deploys the image to a VPS, updates the database (for main branch), updates the PR description with a status badge and URL, and cleans up containers and images on closed PRs.

---------

Co-authored-by: Joshua Jeschek <joshua@jeschek.eu>